Move dashboard to menubar
There is still "no way back."  I was able to actually get it to open a
second window by crashing the application, but that's not sustainable
long-term.

Instead it's a one way move and I don't pretend to be able to bring you
back.  It's also enabled across all environments, because it it owned by
the operator, not me.

THis may disappear entirely in due time, but at least it's off the page.