Add simple alias support
This opens up a whole range of new challenges and opportnities.  For
example, I should be able to use the `~` menu item ON an alias to also
take it.  And there should be more kinds of identity management.

It's built to expand from this current simple working state.

Kinda neat that you can reference where you were when you decided to
apply the alias, if you like.