Add first preferences
As the eldest adjustable, `:iconset` goes first

Then `:identity` and `:clump_id` so that we can entirely drop `Baby`
config.

THere are still some issues wrt the "public/net" identity and the
internal identity, but they will align on restart.  Plus I can fix it
later.

It's much more difficult to handle the Baobab config as a library.
Right now I am letting `Baby` write to it, but this is unsatisfying and
wrong if we want to be able to explore multiple Bamboo clumps.