Add IdentityManager
This required some changes to the `view` concept in the main lve page.
They are better overall with some still obvious warts.

Also have hit a nasty impedence mismatch.  I cannot get a file save
dialog out of a webview.  I could, alternatively, save directly from the
BEAM, but this would break the immersion and lead to confusion.

Instead as it is doing my head in, I leave it in this half-broken state
and hope for inspiration.