Impelement shown/unshown
This distracted me from my intent for the day, but also from my medical
issues.  So there's that.

The design needs work, but I continue to not focus thereon until I can
get the designated designer to do a complete overhaul.

Some of the code is still misplaced, also.  That's just pure laziness.