Adjust all kinds of things for on-boarding ease
I'm not super-excited about all of the compromises here.  I am, however,
pretty excited to have a vaguely reasonable starting point from
an empty application directory.

Lots of niggles, but the core is all here!

The view maintenace comes out of a search for bugs in `Baby` connections.
I was hoping to put it off until a view stack was built for history nav.

I think there is contention if multiple connections are hitting the
`dets` table at the same time.  So one of them might crash.  This crash
annoys the `LiveView` and causes a remount.  (I think this is the behaviour
which I should fix.)

Regardless, it's less impactful if we let it crash and try to maintain
some semblance of what they were looking at. There's still a non-zero
chance that they are on an entires view and we don't have the entry
state.  I can live with thats, for now.

This also brings more clarity to the fact that `identity` awlays
contains a key.  We'll look up the name in Baobab if we need it. We need
it when we need to use the secret key which we will decidedly not be
keeping in the application state.