Add authors route
This is mostly to make it so that I don't need the pesudo-entry forever.
It's pretty Catenary-specific and not so great for an alternative
ecosystem.

This replacement isn't a bowl of cherries, either, to be fair.

In any case, I intend to make it easy to mention people, but I don't
include that here.

I don't like notifications which I suspect is the next step in mentions.
I reckon I can have a route for one to check if one cares.

I also don't like rewriting people's stuff, so I could want soe kind of
control to inject the "correct" Markdown straight-away rather than
magically changing it.

This probably requires some JS and I don't feel up to it.