Make clump_id change kinda work
It doesn't really work, but it kinda looks like it does.

This gives something on which to build for making it actually work

At a minimum need to add restarting the Baby server.  Also the `entry`
and `tag` states.  Probably the `store_hash`, but that may resolve
itself.

Sure to get it wrong early.