Maintain per-clump `:shown` state
There might be an intersection of idenitites wherein the entry
references "point" to the same places, but reference different contents.
THe most likely place for this to happen is our own stuff which we have
nearly surely been shown.  Still, it's best to get it close to correct
now instead of try to remember this and figure it out later.