Add clump compaction
This is a blunt-froce tool.

It's important that we don't compact our own logs except as an explicit
operation. We store data that we tell ourselves (like aliases).
We are also the ultimate source of truth for logs where we control the
keys.  We shouldn't expect others to hold stuff that we ourselves do
not.