"State updates" to Navigation module
This is a minor change which still feels a bit like mixing
responsibilities.  On the other hand, it's working off the `socket` so
that ship has sailed.

In any case, this ought to allow for a navigation update which maintains
a stack of `shown` entries.  I'm still not super-keen on the idea but my
"navigation sucks" from user reports.