Fix forward stack maintenance
When we take a step forward in a new direction, we need to forget about
forward progress on other paths.  This really shows the deficiencies of
linear navgation.  But that's my standard rant.

In any case, there's also a minor refactor including some better naming.
It's still not exactly right, but I lack the imagination to make it so.