Drop UI loop entirely
It spent most of its time doing nothing.  Now we're more event-driven
and reactive, I guess.

Basically, we need to wait for something to say something has happened
and then we reflect it in the UI.

This is less cool for things like "watching unshown stuff stream in",
but that's so niche that I'm probably the only one to have ever done it.

It's easy to "ake things happen, if one knows there is improper state.
Just "do", well, anything and  it will update the state.

I hate the web, but this seems ok-ish for now.