Add useless avatar upload
There is minimal ability to display and explore the images, but I think
a unified image explorer is still in order.

Despite it claiming to be an avatar upload, it's really just a genric
image uploader.  It would we WAY cooler if I could get a file system
listing control without the "upload" nonsense.

It may be that image adding will move to navigation and avatar selection
will be done from an image display page.  I don't have all of the
controls worked out properly.