Precompute tag explorer
This has some minor issues with correctness of shown value and extra
work during computation.

However, for now it will resolve a complaint in clumps wherein the
tagging system is heavily used for navigation.

In the future I will improve both of the above (perhaps by not noting
unshown and building the display versionin parallel, for example)

There is also coming to be a need for pagination.  All of these will be
considered together in the next iteration.  For now, this should assuage
user concerns.