Improve social graph blocking nav
This likely should have been in two commits. The move and the (minor)
update thereto.  I just got annoyed because they weren't together after
fixing.

There was an extraneous `|` character that I don't care to bisect to
find.  It was bad typing for sure.

This also make it possible to block from a profile which should be
expected behaviour.