Improve oasis list indexing with incremntals
I want the state to be deterministic, but I don't want to sort on each
access. I'll fix that bit later.

Next up is ensuring that :graph is doing the right thing

I think it is because I've always known it had to start at the
beginning.

Others properly use their tables to hold randomly accessible values, so
they should be fine.