Massively simplify connection handling
We now properly let `Baby` handle all of the monitoring we just
ask for status. This also reduces some of the work done here and may
improve asscoaited UI mushiness.

In the long term, oases will become a less pivotal portion of the
connections as we add:

- BLE
- mDNS
- Radio BaBa

... and potentially other (better) connection types

We'll eventually be able to query the running processes to generate a
report. Maybe it's the short id and BAMBs sent/recvd.  I am not sure
yet.

As an aside, we'll also soon be able to add more `WANT` and `HAVE`
messages to an extant connection. This will be cool if, for example,
you're connected and hit a missing entry due to log compaction or
partial replication.